+++ src/mizdb_watchlist/static/mizdb_watchlist/js/watchlist.js
@@ -1,4 +1,26 @@
 (() => {
+  function getCSRFToken () {
+    const csrfInput = document.querySelector('[name=csrfmiddlewaretoken]')
+    if (csrfInput) {
+      return csrfInput.value
+    } else {
+      // Get token from cookie:
+      // https://docs.djangoproject.com/en/5.0/howto/csrf/#acquiring-the-token-if-csrf-use-sessions-and-csrf-cookie-httponly-are-false
+      if (document.cookie && document.cookie !== '') {
+        const name = 'csrftoken'
+        const cookies = document.cookie.split(';')
+        for (let i = 0; i < cookies.length; i++) {
+          const cookie = cookies[i].trim()
+          // Does this cookie string begin with the name we want?
+          if (cookie.substring(0, name.length + 1) === (name + '=')) {
+            return decodeURIComponent(cookie.substring(name.length + 1))
+          }
+        }
+      }
+    }
+    throw new Error('No CSRF Token set.')
+  }
+
   function getRequest (btn) {
     const form = new FormData()
     form.append('object_id', btn.dataset.objectId)
@@ -6,7 +28,7 @@
     return new Request(btn.dataset.url, {
       method: 'POST',
       headers: {
-        'X-CSRFToken': document.querySelector('[name=csrfmiddlewaretoken]').value
+        'X-CSRFToken': getCSRFToken()
       },
       body: form,
       mode: 'same-origin'