+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [unreleased]
+
+- get CSRF value from cookie if not using token
+
 ## 0.1.8 (2024-04-10)
 
 - tweak watchlist overview template and CSS